move: 패키지 이동에 영향을 받는 파일 수정
Related to: #32
