+++ bank-application/dev-bank-pay/src/test/java/org/bank/pay/core/family/FamilyEventFacadeTest.java
@@ -3,12 +3,13 @@ package org.bank.pay.core.family;
 import org.assertj.core.api.Assertions;
 import org.bank.core.auth.AuthClaims;
 import org.bank.core.payment.Product;
-import org.bank.pay.core.cash.UserRegisterTask;
 import org.bank.pay.core.domain.familly.Family;
 import org.bank.pay.core.domain.familly.FamilyService;
 import org.bank.pay.core.domain.familly.MemberClaims;
 import org.bank.pay.core.domain.familly.repository.FamilyReader;
-import org.bank.pay.core.domain.onwer.OwnerClaims;
+import org.bank.pay.core.domain.owner.OwnerClaims;
+import org.bank.pay.core.domain.owner.PayOwner;
+import org.bank.pay.core.domain.owner.repository.PayOwnerStore;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -39,9 +40,9 @@ class FamilyEventFacadeTest {
     @Autowired
     private FamilyService familyService;
     @Autowired
-    private UserRegisterTask userRegisterTask;
-    @Autowired
     private FamilyReader familyReader;
+    @Autowired
+    private PayOwnerStore payOwnerStore;
 
     private final AuthClaims leader = new OwnerClaims("fixture-id", "fixture", "fixture@email.com");
     private final MemberClaims follower = new MemberClaims("follower", "follower", "follower@email.com");
@@ -50,7 +51,8 @@ class FamilyEventFacadeTest {
 
     @BeforeAll
     public void setup() {
-        userRegisterTask.initialize(leader);
+
+        payOwnerStore.save(new PayOwner(leader));
         familyService.createFamily(leader);
 
         MemberClaims familyLeader = MemberClaims.of(leader);
@@ -62,7 +64,7 @@ class FamilyEventFacadeTest {
     void 새로운_멤버를_초대합니다() {
 
         Assertions.assertThat(family).isNotNull();
-        Assertions.assertThatCode(() -> familyEventFacade.inviteMember(family.getFamilyId(), leader, follower))
+        Assertions.assertThatCode(() -> familyEventFacade.invite(family.getFamilyId(), leader, follower.getUserid()))
                         .doesNotThrowAnyException();
     }
 
@@ -74,7 +76,7 @@ class FamilyEventFacadeTest {
         familyService.addMemberToFamily(family.getFamilyId(), participant);
 
         Assertions.assertThat(family).isNotNull();
-        Assertions.assertThatCode(() -> familyEventFacade.requestPayment(family.getFamilyId(), participant.getUserid(), new Product("예제 상품", 1000, 1)))
+        Assertions.assertThatCode(() -> familyEventFacade.purchaseProductToFamilyLeader(family.getFamilyId(), participant.getUserid(), new Product("예제 상품", 1000, 1)))
                 .doesNotThrowAnyException();
     }
 
@@ -86,7 +88,7 @@ class FamilyEventFacadeTest {
         familyService.addMemberToFamily(family.getFamilyId(), participant);
 
         Assertions.assertThat(family).isNotNull();
-        Assertions.assertThatCode(() -> familyEventFacade.requestPayment(family.getFamilyId(), participant.getUserid(), new Product("예제 상품", 1000, 1)))
+        Assertions.assertThatCode(() -> familyEventFacade.purchaseProductToFamilyLeader(family.getFamilyId(), participant.getUserid(), new Product("예제 상품", 1000, 1)))
                 .doesNotThrowAnyException();
     }
 }
